+++ src/coreaudiodevice.rs
@@ -442,7 +442,7 @@ impl PlaybackDevice for CoreaudioPlaybackDevice {
                                     timer.restart();
                                     buffer_avg.restart();
                                     debug!(
-                                        "Current buffer level {}, set capture rate to {}%",
+                                        "Current buffer level {:.1}, set capture rate to {:.4}%",
                                         av_delay,
                                         100.0 * speed
                                     );