+++ src/wasapidevice.rs
@@ -692,7 +692,7 @@ impl PlaybackDevice for WasapiPlaybackDevice {
                                     timer.restart();
                                     buffer_avg.restart();
                                     debug!(
-                                        "Current buffer level {}, set capture rate to {}%",
+                                        "Current buffer level {:.1}, set capture rate to {:.4}%",
                                         av_delay,
                                         100.0 * speed
                                     );